requirements.txt
torch==2.6.0
transformers==4.46.3
tokenizers==0.20.3
einops
addict 
easydict
Pillow
torchvision==0.21.0
# pip install flash-attn==2.7.3 --no-build-isolation
https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.3/flash_attn-2.7.3+cu12torch2.6cxx11abiFALSE-cp312-cp312-linux_x86_64.whl